Removed the received map.
I've removed the received map in the QueueCache class as I'm not sure
what it's for and it means that I can't delete messages in a mock queue
until I have called receive. I see why as this would be the normal logic
t get the receipt handles but in our use case, I don't need it and I
don't think it's necessary anyway. If I'm wrong, I'll add it back.

I've also derived the receipt handle from the body. This means that I
can tell what the receipt handle will be without having to query the
messages.